Use json-spec-validator which compiles with nuitka
@pbelmann In relation to bioboxes/rfc#86 I have had to switch from the
json-schema to use json-schema-validator. Unfortunately json-schema is
not compatible with nuitka as it loads a json file while it is being
used. This json file is no longer present after the program has been
compiled and so will not work.

I prefer json-schema as I find the error messages clearer. If necessary,
in future, we can fork or send a pull request to make this version
compatible with nuitka.

Signed-off-by: Michael Barton <mail@michaelbarton.me.uk>